conscience_service/src/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.0.10
django-cors-headers==4.6.0
django-ninja==1.3.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.1.0
packaging==24.2
propcache==0.2.1
psycopg==3.2.3
psycopg-binary==3.2.3
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2
yarl==1.18.3
rav
httpx 